Develop -> main push

+++ src/main/java/org/likelion/likelion_12th_team05/comment/appication/CommentService.java
@@ -1,5 +1,6 @@
 package org.likelion.likelion_12th_team05.comment.appication;
 
+import org.likelion.likelion_12th_team05.comment.api.request.CommentSaveReqDto;
 import org.likelion.likelion_12th_team05.comment.api.request.CommentUpdateReqDto;
 import org.likelion.likelion_12th_team05.comment.api.response.CommentInfoResDto;
 import org.likelion.likelion_12th_team05.comment.domain.Comment;
@@ -7,8 +8,6 @@ import org.likelion.likelion_12th_team05.comment.domain.CommentRepository;
 import org.likelion.likelion_12th_team05.common.EntityFinder;
 import org.likelion.likelion_12th_team05.common.error.ErrorCode;
 import org.likelion.likelion_12th_team05.common.exception.CustomException;
-import org.likelion.likelion_12th_team05.curation.api.dto.request.CurationSaveReqDto;
-import org.likelion.likelion_12th_team05.curation.api.dto.response.CurationInfoResDto;
 import org.likelion.likelion_12th_team05.curation.domain.Curation;
 import org.likelion.likelion_12th_team05.curation.domain.repository.CurationRepository;
 import org.likelion.likelion_12th_team05.user.domain.User;
@@ -32,23 +31,24 @@ public class CommentService {
     }
 
     @Transactional
-    public CurationInfoResDto commentSave(CurationSaveReqDto curationSaveReqDto, Principal principal) {
+    public CommentInfoResDto commentSave(Long curationId, CommentSaveReqDto commentSaveReq, Principal principal) {
+        Curation curation = getCurationById(curationId);    // 큐레이션 엔티티 조회 <- 큐레이션의 댓글 카운트 설정을 위함
         String email = principal.getName();
         User user = getUserByEmail(email);
 
-        Curation curation = curationSaveReqDto.toEntity(user);
-        curation.setCommentCount(curation.getCommentCount() + 1);
-        curationRepository.save(curation);
+        Comment comment = commentSaveReq.toEntity(user, curation);    // toEntity로 user 엔티티 정보와 함께 comment 객체 생성
+        curation.setCommentCount(curation.getCommentCount() + 1);   // curation의 commentCount를 get 해와 + 1 한 다음에 setCommentCount로 세팅해줌
+        commentRepository.save(comment);    // 댓글 저장
 
-        return CurationInfoResDto.from(curation);
+        return CommentInfoResDto.from(comment); // 댓글 정보 리턴
     }
 
     @Transactional
-    public CommentInfoResDto commentUpdate(Long curationId, CommentUpdateReqDto commentUpdateReqDto, Principal principal) {
+    public CommentInfoResDto commentUpdate(Long commentId, CommentUpdateReqDto commentUpdateReqDto, Principal principal) {
         String email = principal.getName();
         User user = getUserByEmail(email);
 
-        Comment comment = getCommentById(curationId);
+        Comment comment = getCommentById(commentId);
 
         String commentUserEmail = comment.getUser().getEmail();
         if (!email.equals(commentUserEmail)) {
@@ -64,12 +64,18 @@ public class CommentService {
     @Transactional
     public void commentDelete(Long commentId, Principal principal) {
         String email = principal.getName();
-        User user = getUserByEmail(email);
 
-        Comment comment = commentRepository.findById(commentId).orElseThrow(
-                () -> new CustomException(ErrorCode.CURATIONS_NOT_FOUND_EXCEPTION,
-                        ErrorCode.USER_NOT_FOUND_EXCEPTION.getMessage()));
+        Comment comment = getCommentById(commentId);
 
+        String commentUserEmail = comment.getUser().getEmail();
+        if (!email.equals(commentUserEmail)) {
+            throw new CustomException(ErrorCode.NO_AUTHORIZATION_EXCEPTION,
+                    ErrorCode.NO_AUTHORIZATION_EXCEPTION.getMessage());
+        }
+
+        Curation curation = comment.getCuration();
+        if (curation != null)
+            curation.setCommentCount(curation.getCommentCount() - 1);
         commentRepository.delete(comment);
     }
 
@@ -81,5 +87,10 @@ public class CommentService {
     private User getUserByEmail(String email) {
         return EntityFinder.findByEmailOrThrow
                 (userRepository.findByEmail(email), ErrorCode.USER_NOT_FOUND_EXCEPTION);
+    }
+
+    private Curation getCurationById(Long curationId) {
+        return EntityFinder.findByIdOrThrow(curationRepository.findById(curationId)
+                , ErrorCode.CURATIONS_NOT_FOUND_EXCEPTION);
     }
 }
